[Data] Move env_float, env_integer, and env_bool to ray._common
