+++ python/ray/_common/utils.py
@@ -3,6 +3,7 @@ import binascii
 import errno
 import importlib
 import inspect
+import logging
 import os
 import random
 import string
@@ -15,6 +16,45 @@ from types import ModuleType
 from typing import Any, Coroutine, Dict, Optional, Tuple
 
 import psutil
+
+logger = logging.getLogger(__name__)
+
+
+def env_integer(key, default):
+    if key in os.environ:
+        value = os.environ[key]
+        try:
+            return int(value)
+        except ValueError:
+            logger.debug(
+                f"Found {key} in environment, but value must "
+                f"be an integer. Got: {value}. Returning "
+                f"provided default {default}."
+            )
+            return default
+    return default
+
+
+def env_float(key, default):
+    if key in os.environ:
+        value = os.environ[key]
+        try:
+            return float(value)
+        except ValueError:
+            logger.debug(
+                f"Found {key} in environment, but value must "
+                f"be a float. Got: {value}. Returning "
+                f"provided default {default}."
+            )
+            return default
+    return default
+
+
+def env_bool(key, default):
+    if key in os.environ:
+        val = os.environ[key].lower()
+        return val == "true" or val == "1"
+    return default
 
 
 def import_module_and_attr(
