+++ python/ray/data/__init__.py
@@ -106,7 +106,7 @@ try:
     if pyarrow_version is None or pyarrow_version >= parse_version("21.0.0"):
         pass
     else:
-        from ray._private.ray_constants import env_bool
+        from ray._common.utils import env_bool
 
         RAY_DATA_AUTOLOAD_PYEXTENSIONTYPE = env_bool(
             "RAY_DATA_AUTOLOAD_PYEXTENSIONTYPE", False
