+++ python/ray/data/_internal/planner/plan_udf_map_op.py
@@ -28,7 +28,7 @@ import pyarrow as pa
 
 import ray
 from ray._common.utils import get_or_create_event_loop
-from ray._private.ray_constants import env_integer
+from ray._common.utils import env_integer
 from ray.data._internal.compute import ActorPoolStrategy, ComputeStrategy, get_compute
 from ray.data._internal.execution.interfaces import PhysicalOperator
 from ray.data._internal.execution.interfaces.task_context import TaskContext
